@@ -23,32 +23,32 @@
     "deploy:prod": "vercel deploy --prod"
   },
   "dependencies": {
-    "@fastify/cors": "^10.0.2",
+    "@fastify/cors": "^11.2.0",
     "@types/ws": "^8.18.1",
-    "chalk": "^5.4.1",
+    "chalk": "^5.6.2",
     "colorthief": "^2.6.0",
     "crypto-js": "^4.2.0",
-    "dotenv": "^16.4.7",
-    "fastify": "^5.2.1",
-    "form-data": "^4.0.2",
-    "ioredis": "^5.6.0",
-    "nanoid": "^5.1.0",
-    "puppeteer": "^24.17.0",
+    "dotenv": "^17.2.3",
+    "fastify": "^5.7.4",
+    "form-data": "^4.0.5",
+    "ioredis": "^5.9.2",
+    "nanoid": "^5.1.6",
+    "puppeteer": "^24.36.1",
     "tslib": "~2.8.1",
-    "ws": "^8.18.2",
+    "ws": "^8.19.0",
     "xmldom": "^0.6.0"
   },
   "devDependencies": {
-    "@biomejs/biome": "1.9.4",
-    "@types/jest": "~29.5",
-    "@types/node": "~22",
-    "cross-env": "^7.0.3",
+    "@biomejs/biome": "2.3.14",
+    "@types/jest": "~30.0",
+    "@types/node": "~25",
+    "cross-env": "^10.1.0",
     "husky": "^9.1.7",
-    "lint-staged": "^15.4.3",
-    "nodemon": "^3.1.9",
-    "rimraf": "~6.0",
-    "ts-jest": "~29.2",
-    "tsx": "^4.19.3",
-    "typescript": "~5.7"
+    "lint-staged": "^16.2.7",
+    "nodemon": "^3.1.11",
+    "rimraf": "~6.1",
+    "ts-jest": "~29.4",
+    "tsx": "^4.21.0",
+    "typescript": "~5.9"
   }
 }
