@@ -1,73 +1,75 @@
-import { isValidArray } from '../helpers/common';
-import { gaanaDataConfigs } from '../helpers/dataConfig';
-import { dataExtractor } from '../helpers/dataExtractor';
-const ColorThief = require('colorthief');
+import {
+  mapGaanaAlbum,
+  mapGaanaArtist,
+  mapGaanaPlaylist,
+  mapGaanaSong,
+} from '../mappers/gaana.mapper';
+import { Album, Artist, Playlist, Song } from '../types';
 
-async function getPalettes(data) {
-  const dataWithPalette = [];
+export const gaanaDataSanitizer = (data: any): Song | Album | Playlist | Artist | any => {
+  const type = data.entity_type;
 
-  for (const item of data) {
-    try {
-      const palette = await ColorThief.getPalette(item.image, 5);
-      dataWithPalette.push({ ...item, palette });
-    } catch (error) {
-      console.error(`Error processing ${item}:`, error);
-    }
-  }
-  return dataWithPalette;
-}
+  if (type === 'TR') return mapGaanaSong(data);
+  if (type === 'AL') return mapGaanaAlbum(data);
+  if (type === 'PL') return mapGaanaPlaylist(data);
+  if (type === 'AR') return mapGaanaArtist(data);
 
-const radioDataMapper = (data: any, page: number) => {
-  if (data.status === 0) {
-    return [];
-  }
-
-  if (data.entities && isValidArray(data.entities)) {
-    const sanitizedData = data.entities.map((radioData) => {
-      const extractedData: any = {};
-      for (let key in gaanaDataConfigs.radio) {
-        extractedData[key] = dataExtractor(radioData, gaanaDataConfigs.radio[key]);
-        extractedData.source = 'gaana';
-      }
-      return extractedData;
-    });
-
-    return {
-      list: sanitizedData,
-      page: Number(page),
-      count: data?.count,
-    };
-  }
+  return data;
 };
 
-const songDataMapper = (tracks) => {
-  const sanitizedData = tracks.map((track) => {
-    const extractedData: any = {};
-    for (let key in gaanaDataConfigs.song) {
-      extractedData[key] = dataExtractor(track, gaanaDataConfigs.song[key]);
-      extractedData.source = 'gaana';
+export const gaanaSearchMapper = (data: any) => {
+  const gr = data.gr || [];
+  const sanitizedData = [];
+
+  gr.forEach((group: any) => {
+    const heading = group.stxt;
+    const items = (group.gd || []).map((item: any) => {
+      // Normalizing search item to match our internal structure
+      const normalizedItem = {
+        entity_id: item.id,
+        name: item.ti,
+        seokey: item.seo,
+        artwork: item.aw,
+        entity_type:
+          item.ty === 'Track'
+            ? 'TR'
+            : item.ty === 'Album'
+              ? 'AL'
+              : item.ty === 'Playlist'
+                ? 'PL'
+                : item.ty === 'Artist'
+                  ? 'AR'
+                  : item.ty,
+        language: item.language,
+        entity_info: [
+          {
+            key: 'artist',
+            value: (item.sti || '').split(',').map((name: string) => ({ name: name.trim() })),
+          },
+          { key: 'album', value: [{ name: item.sti }] },
+        ],
+      };
+      return gaanaDataSanitizer(normalizedItem);
+    });
+
+    if (items.length > 0) {
+      sanitizedData.push({
+        heading,
+        data: items,
+        source: 'gaana',
+      });
     }
-
-    extractedData.artists = extractedData.artists.map((artist: any) => ({
-      id: artist.artist_id,
-      name: artist.name,
-      type: 'artist',
-      key: artist.seokey,
-    }));
-
-    extractedData.genres = extractedData.genres.map((genre: any) => ({
-      id: genre.genre,
-      name: genre.name,
-    }));
-
-    return extractedData;
   });
+
   return sanitizedData;
 };
 
-const radioDetailMapper = async (data: any) => {
-  const sanitizedSongData = songDataMapper(data.tracks);
-  return sanitizedSongData;
+export const gaanaHomeMapper = (data: any): any[] => {
+  return (data.home || []).map((section: any) => ({
+    heading: section.ga_header || section.entity_description,
+    source: 'gaana',
+    url: section.url,
+    seokey: section.seokey_url,
+    viewType: section.view_type,
+  }));
 };
-
-export { radioDataMapper, radioDetailMapper };
