@@ -115,68 +115,4 @@ const saavnDataConfigs = {
   },
 };
 
-const gaanaDataConfigs = {
-  radio: {
-    language: 'language',
-    key: 'seokey',
-    name: 'name',
-    image: 'artwork',
-    id: 'entity_id',
-    type: 'entity_type',
-    favorites: 'favorite_count',
-    count: '',
-  },
-  song: {
-    id: 'track_id',
-    title: 'track_title',
-    subtitle: 'album_title',
-    images: 'artwork_large',
-    image: 'artwork',
-    language: 'language',
-    album: 'album_title',
-    albumId: 'album_id',
-    duration: 'duration',
-    releaseDate: 'release_date',
-    artists: 'artist',
-    key: 'seokey',
-    albumKey: 'albumseokey',
-    genres: 'gener',
-  },
-};
-
-const ytmDataConfigs = {
-  songConfig: {
-    id: 'videoId',
-    title: 'title',
-    subtitle: 'subtitle',
-    images: 'thumbnails.1.url',
-    image: 'thumbnails.1.url',
-    album: 'album.name',
-    albumId: 'album.id',
-    artists: 'artists',
-    isExplicit: 'isExplicit',
-    duration: 'duration',
-  },
-  albumOrPlaylistConfig: {
-    id: 'browseId',
-    radioId: 'radioId',
-    playlistId: ['audioPlaylistId', 'playlistId'],
-    description: 'description',
-    title: ['title', 'name'],
-    subtitle: 'subtitle',
-    type: 'type',
-    token: 'perma_url',
-    desc: 'description',
-    images: 'thumbnails.1.url',
-    year: 'year',
-    isExplicit: 'isExplicit',
-    artist: 'artists',
-  },
-  contentConfig: {
-    id: 'id',
-    title: 'title',
-    contents: 'contents',
-  },
-};
-
-export { gaanaDataConfigs, saavnDataConfigs, ytmDataConfigs };
+export { saavnDataConfigs };
