@@ -3,12 +3,14 @@ import { cache } from '../app';
 import { config } from '../config/config';
 import { isValidArray } from '../helpers/common';
 import { fetchGet } from '../helpers/http';
+import { sendError, sendSuccess } from '../utils/response';
 import {
   albumDataMapper,
   artistDataMapper,
   autoCompleteDataMapper,
   homeDataMapper,
   modulesDataMapper,
+  playlistDataMapper,
   recommendedAlbumDataMapper,
   songDataSanitizer,
   songsDetailsMapper,
@@ -20,6 +22,7 @@ type SaavnRequest = FastifyRequest<{
   Querystring: {
     languages: string;
     language: string;
+    lang: string;
     page: number;
     count: number;
     name: string;
@@ -44,53 +47,80 @@ const params = {
   ctx: config.saavn.ctx,
 };
 
+const saavnFetch = async <T>(url: string, options: any = {}) => {
+  const languages = options.lang || 'hindi,english';
+  // Saavn L cookie uses comma separated lowercase languages, usually encoded
+  const formattedLangs = languages
+    .split(',')
+    .map((l: string) => l.trim().toLowerCase())
+    .join(',');
+
+  const headers = {
+    ...options.headers,
+    cookie: `B=b7984e01db109802c47c5178fac7badc; CT=MjA2Mjg5ODMw; _pl=web6dot0-; DL=english; L=${encodeURIComponent(formattedLangs)}; geo=49.207.50.17%2CIN%2CKarnataka%2CBengaluru%2C562130; mm_latlong=12.9753%2C77.591; CH=G03%2CA07%2CO00%2CL03.; gdpr_acceptance=true`,
+  };
+
+  return fetchGet<T>(url, { ...options, headers });
+};
+
 const homeController = async (req: SaavnRequest, res: FastifyReply) => {
-  const languages = req.query.languages;
+  const languages = req.query.lang || req.query.languages;
   const url = `${config.saavn.baseUrl}?__call=${config.saavn.endpoint.modules.home}`;
   const key = `home_${languages}`;
 
   const cacheData = await cache.get(key);
   if (cacheData) {
-    res.code(cacheData.code).send({
-      ...cacheData,
-    });
-  } else {
-    const { data, code, error, message } = await fetchGet(url, {
-      params: {
-        ...params,
-      },
-      headers: {
-        cookie: `L=${languages || 'hindi,punjabi'}; gdpr_acceptance=true; DL=english`,
-      },
-    });
-
-    const sanitizedData = homeDataMapper(data);
-    cache.set(key, { code, message, data: sanitizedData, error });
-
-    res.code(code).send({ code, message, data: sanitizedData, error });
+    return res.code(cacheData.code || 200).send(cacheData);
   }
+
+  const { data, code, error, message } = await saavnFetch<any>(url, {
+    params: {
+      ...params,
+    },
+    lang: languages,
+  });
+
+  if (error) {
+    return sendError(res, message || 'Failed to fetch home data', error, code);
+  }
+
+  const sanitizedData = homeDataMapper(data);
+  const response = {
+    status: 'success',
+    message: message || 'Home data fetched successfully',
+    data: sanitizedData,
+    error: null,
+    source: 'saavn',
+  };
+
+  cache.set(key, { ...response, code });
+  return res.code(code).send(response);
 };
 
 const modulesController = async (req: SaavnRequest, res: FastifyReply) => {
-  const languages = req.query.languages;
+  const languages = req.query.lang || req.query.languages;
   const url = `${config.saavn.baseUrl}`;
-  const { data, code, error, message } = await fetchGet(url, {
+  const { data, code, error, message } = await saavnFetch<any>(url, {
     params: {
       ...params,
       __call: config.saavn.endpoint.modules.browse_modules,
     },
-    headers: {
-      cookie: `L=${languages || 'hindi'}; gdpr_acceptance=true; DL=english`,
-    },
+    lang: languages,
   });
+
+  if (error) {
+    return sendError(res, message || 'Failed to fetch modules', error, code);
+  }
+
   const sanitizedData = modulesDataMapper(data);
-  res.code(code).send({ code, message, data: sanitizedData, error });
+  return sendSuccess(res, sanitizedData, message, 'saavn', code);
 };
 
 const albumRecommendationController = async (req: SaavnRequest, res: FastifyReply) => {
   const { albumId } = req.params;
+  const languages = req.query.lang || req.query.languages;
   const url = `${config.saavn.baseUrl}`;
-  const { data, code, error, message } = await fetchGet(url, {
+  const { data, code, error, message } = await saavnFetch<any>(url, {
     params: {
       __call: config.saavn.endpoint.album.recommended,
       albumid: albumId,
@@ -98,97 +128,107 @@ const albumRecommendationController = async (req: SaavnRequest, res: FastifyRepl
       includeMetaTags: 0,
       ...params,
     },
+    lang: languages,
   });
-  const sanitizedData = recommendedAlbumDataMapper(data);
-  res.code(code).send({ code, message, data: sanitizedData, error });
+
+  if (error) {
+    return sendError(res, message || 'Failed to fetch album recommendations', error, code);
+  }
+
+  const sanitizedData = recommendedAlbumDataMapper(data as any[]);
+  return sendSuccess(res, sanitizedData, message, 'saavn', code);
 };
 
 const topAlbumsOfYearController = async (req: SaavnRequest, res: FastifyReply) => {
   const { year } = req.params;
+  const languages = req.query.lang || req.query.languages;
   const url = `${config.saavn.baseUrl}`;
-  const { data, code, error, message } = await fetchGet(url, {
+  const { data, code, error, message } = await saavnFetch<any>(url, {
     params: {
       __call: config.saavn.endpoint.album.top_albums_by_year,
       album_year: year,
       ...params,
     },
+    lang: languages,
   });
 
-  const sanitizedData = recommendedAlbumDataMapper(data);
-  res.code(code).send({ code, message, data: sanitizedData, error });
+  if (error) {
+    return sendError(res, message || 'Failed to fetch top albums', error, code);
+  }
+
+  const sanitizedData = recommendedAlbumDataMapper(data as any[]);
+  return sendSuccess(res, sanitizedData, message, 'saavn', code);
 };
 
 const albumController = async (req: SaavnRequest, res: FastifyReply) => {
   try {
     const { albumId } = req.params;
-    const cacheData = await cache.get(albumId);
+    const languages = req.query.lang || req.query.languages;
+    const cacheData = await cache.get(`${albumId}_${languages}`);
     if (cacheData) {
-      res.code(cacheData.code).send({
-        ...cacheData,
-      });
-    } else {
-      const url = `${config.saavn.baseUrl}`;
-      const { data, code, error, message } = await fetchGet(url, {
-        params: {
-          __call: config.saavn.endpoint.album.token,
-          token: albumId,
-          type: 'album',
-          includeMetaTags: 0,
-          ...params,
-        },
-      });
-      const sanitizedData: any = albumDataMapper(data);
-      const promiseArr = sanitizedData.modules.map((d) => {
-        return {
-          title: d.heading,
-          promise: fetchGet(url, {
-            params: {
-              __call: d.endpoint,
-              albumid: d.albumId,
-              album_year: d.year,
-              type: d.type,
-              language: d.language,
-              ...params,
-            },
-          }),
-        };
-      });
-      const results = await Promise.allSettled(promiseArr.map((item) => item.promise));
-
-      const sections = results
-        .map((result: any, index) =>
-          isValidArray(result.value.data)
-            ? {
-                heading: promiseArr[index].title,
-                data: isValidArray(result.value.data)
-                  ? recommendedAlbumDataMapper(result.value.data)
-                  : undefined,
-              }
-            : null,
-        )
-        .filter((d: any) => d != null);
-      res.code(code).send({
-        code,
-        message,
-        data: { album: sanitizedData.album, sections, data },
-        error,
-      });
-
-      //setting up valkey cache data
-      cache.set(albumId, {
-        code,
-        message,
-        data: { album: sanitizedData.album, sections, data },
-        error,
-      });
+      return res.code(cacheData.code || 200).send(cacheData);
     }
-  } catch (error) {
-    res.code(400).send({
-      data: null,
-      code: 400,
-      message: 'failed to fetch',
-      error,
+
+    const url = `${config.saavn.baseUrl}`;
+    const { data, code, error, message } = await saavnFetch<any>(url, {
+      params: {
+        __call: config.saavn.endpoint.album.token,
+        token: albumId,
+        type: 'album',
+        includeMetaTags: 0,
+        ...params,
+      },
+      lang: languages,
     });
+
+    if (error) {
+      return sendError(res, message || 'Failed to fetch album details', error, code);
+    }
+
+    const sanitizedData: any = albumDataMapper(data);
+    const promiseArr = (sanitizedData.modules || []).map((d: any) => {
+      return {
+        title: d.heading,
+        promise: saavnFetch<any>(url, {
+          params: {
+            __call: d.endpoint,
+            albumid: d.albumId,
+            album_year: d.year,
+            type: d.type,
+            language: d.language,
+            ...params,
+          },
+          lang: languages,
+        }),
+      };
+    });
+    const results = await Promise.allSettled(promiseArr.map((item) => item.promise));
+
+    const sections = results
+      .map((result: any, index) =>
+        isValidArray(result.value?.data)
+          ? {
+              heading: promiseArr[index].title,
+              data: recommendedAlbumDataMapper(result.value.data),
+              source: 'saavn',
+            }
+          : null,
+      )
+      .filter((d: any) => d != null);
+
+    const finalData = { album: sanitizedData.album || sanitizedData, sections };
+    const response = {
+      status: 'success',
+      message: message || 'Album details fetched successfully',
+      data: finalData,
+      error: null,
+      source: 'saavn',
+    };
+
+    cache.set(`${albumId}_${languages}`, { ...response, code });
+    return res.code(code).send(response);
+  } catch (error) {
+    return sendError(res, 'Failed to fetch album details', error);
   }
 };
 
@@ -196,8 +236,9 @@ const playlistController = async (req: SaavnRequest, res: FastifyReply) => {
   try {
     const { playlistId } = req.params;
     const { page = 0, count = 50 } = req.query;
+    const languages = req.query.lang || req.query.languages;
     const url = `${config.saavn.baseUrl}`;
-    const { data, code, error, message } = await fetchGet(url, {
+    const { data, code, error, message } = await saavnFetch<any>(url, {
       params: {
         __call: config.saavn.endpoint.playlist.token,
         token: playlistId,
@@ -207,12 +248,18 @@ const playlistController = async (req: SaavnRequest, res: FastifyReply) => {
         p: page,
         ...params,
       },
+      lang: languages,
     });
-    const sanitizedData: any = albumDataMapper(data);
-    const promiseArr = sanitizedData.modules.map((d) => {
+
+    if (error) {
+      return sendError(res, message || 'Failed to fetch playlist', error, code);
+    }
+
+    const sanitizedData: any = playlistDataMapper(data);
+    const promiseArr = (sanitizedData.modules || []).map((d: any) => {
       return {
         title: d.heading,
-        promise: fetchGet(url, {
+        promise: saavnFetch<any>(url, {
           params: {
             __call: d.endpoint,
             listid: d.listId,
@@ -220,6 +267,7 @@ const playlistController = async (req: SaavnRequest, res: FastifyReply) => {
             entity_language: d.language,
             ...params,
           },
+          lang: languages,
         }),
       };
     });
@@ -227,69 +275,75 @@ const playlistController = async (req: SaavnRequest, res: FastifyReply) => {
 
     const sections = results
       .map((result: any, index) =>
-        isValidArray(result.value.data)
+        isValidArray(result.value?.data)
           ? {
               heading: promiseArr[index].title,
-              data: isValidArray(result.value.data)
-                ? recommendedAlbumDataMapper(result.value.data)
-                : undefined,
+              data: recommendedAlbumDataMapper(result.value.data),
+              source: 'saavn',
             }
           : null,
       )
       .filter((d: any) => d != null);
-    res.code(code).send({
-      code,
-      message,
-      data: { album: sanitizedData.album, sections },
-      error,
-    });
+
+    const finalData = { playlist: sanitizedData.playlist, sections };
+    return sendSuccess(res, finalData, message, 'saavn', code);
   } catch (error) {
-    res.code(400).send({
-      data: null,
-      code: 400,
-      message: 'failed to fetch',
-      error,
-    });
+    return sendError(res, 'Failed to fetch playlist', error);
   }
 };
 
 const mixController = async (req: SaavnRequest, res: FastifyReply) => {
   const { mixId } = req.params;
+  const languages = req.query.lang || req.query.languages;
   const url = `${config.saavn.baseUrl}`;
-  const { data, code, error, message } = await fetchGet(url, {
+  const { data, code, error, message } = await saavnFetch<any>(url, {
     params: {
       __call: config.saavn.endpoint.playlist.token,
       token: mixId,
       type: 'mix',
-      p: 1, // TODO: Fix this Page
-      n: 100, // TODO: Fix this count
+      p: 1,
+      n: 100,
       includeMetaTags: 0,
       ...params,
     },
+    lang: languages,
   });
-  const sanitizedData = await albumDataMapper(data);
-  res.code(code).send({ code, message, data: sanitizedData, error });
+
+  if (error) {
+    return sendError(res, message || 'Failed to fetch mix', error, code);
+  }
+
+  const sanitizedData = albumDataMapper(data);
+  return sendSuccess(res, sanitizedData, message, 'saavn', code);
 };
 
 const stationController = async (req: SaavnRequest, res: FastifyReply) => {
   const { language, name } = req.query;
+  const languages = req.query.lang || req.query.languages || language;
   const url = `${config.saavn.baseUrl}`;
-  const { data, code, error, message } = await fetchGet(url, {
+  const { data, code, error, message } = await saavnFetch<any>(url, {
     params: {
       __call: config.saavn.endpoint.radio.featured,
-      language: language,
+      language: languages,
       name: name,
       includeMetaTags: 0,
       ...params,
     },
+    lang: languages,
   });
-  res.code(code).send({ code, message, data: data, error });
+
+  if (error) {
+    return sendError(res, message || 'Failed to create station', error, code);
+  }
+
+  return sendSuccess(res, data, message, 'saavn', code);
 };
 
 const stationSongsController = async (req: SaavnRequest, res: FastifyReply) => {
   const { stationId, count } = req.query;
+  const languages = req.query.lang || req.query.languages;
   const url = `${config.saavn.baseUrl}`;
-  const { data, code, error, message } = await fetchGet(url, {
+  const { data, code, error, message } = await saavnFetch<any>(url, {
     params: {
       __call: config.saavn.endpoint.radio.songs,
       stationid: stationId,
@@ -297,35 +351,46 @@ const stationSongsController = async (req: SaavnRequest, res: FastifyReply) => {
       k: count,
       ...params,
     },
+    lang: languages,
   });
+
+  if (error) {
+    return sendError(res, message || 'Failed to fetch station songs', error, code);
+  }
+
   const sanitizedData = await stationSongsMapper(data);
-  res.code(code).send({ code, message, data: sanitizedData, error });
+  return sendSuccess(res, sanitizedData, message, 'saavn', code);
 };
 
 const topSearchController = async (req: SaavnRequest, res: FastifyReply) => {
+  const languages = req.query.lang || req.query.languages;
   const url = `${config.saavn.baseUrl}`;
-  const { data, code, error, message } = await fetchGet(url, {
+  const { data, code, error, message } = await saavnFetch<any>(url, {
     params: {
       __call: config.saavn.endpoint.search.top_search,
       ...params,
     },
+    lang: languages,
   });
 
-  const sortedData = topSearchMapper(data);
+  if (error) {
+    return sendError(res, message || 'Failed to fetch top search', error, code);
+  }
 
-  res.code(code).send({ code, message, data: sortedData, error });
+  const sortedData = topSearchMapper(data);
+  return sendSuccess(res, sortedData, message, 'saavn', code);
 };
 
 const searchController = async (req: SaavnRequest, res: FastifyReply) => {
-  // Explicitly type and normalize the query parameters
   const q = req.query.q as string;
   const type = ((req.query.type as string) || 'all').toLowerCase();
   const page = Number(req.query.page) || 1;
   const count = Number(req.query.count) || 30;
+  const languages = req.query.lang || req.query.languages;
 
-  let url = `${config.saavn.baseUrl}`;
+  const url = `${config.saavn.baseUrl}`;
 
-  const { data, code, error, message } = await fetchGet(url, {
+  const { data, code, error, message } = await saavnFetch<any>(url, {
     params: {
       __call: config.saavn.endpoint.search[type],
       q,
@@ -335,7 +400,12 @@ const searchController = async (req: SaavnRequest, res: FastifyReply) => {
       ...params,
       api_version: ['artist', 'all'].includes(type) ? 3 : 4,
     },
+    lang: languages,
   });
+
+  if (error) {
+    return sendError(res, message || 'Search failed', error, code);
+  }
 
   let sanitizedData;
   try {
@@ -343,7 +413,7 @@ const searchController = async (req: SaavnRequest, res: FastifyReply) => {
       case 'albums':
         sanitizedData = {
           heading: 'Albums',
-          list: (data as any).results.map((d) => albumDataMapper(d)),
+          list: (data as any).results.map((d: any) => albumDataMapper(d)),
           source: 'saavn',
           count: (data as any).total,
         };
@@ -355,12 +425,11 @@ const searchController = async (req: SaavnRequest, res: FastifyReply) => {
           source: 'saavn',
           count: (data as any).total,
         };
-
         break;
       case 'artists':
         sanitizedData = {
           heading: 'Artists',
-          list: (data as any).results.map((d) => albumDataMapper(d)),
+          list: (data as any).results.map((d: any) => artistDataMapper(d)),
           source: 'saavn',
           count: (data as any).total,
         };
@@ -368,7 +437,7 @@ const searchController = async (req: SaavnRequest, res: FastifyReply) => {
       case 'playlists':
         sanitizedData = {
           heading: 'Playlists',
-          list: (data as any).results.map((d) => albumDataMapper(d)),
+          list: (data as any).results.map((d: any) => playlistDataMapper(d)),
           source: 'saavn',
           count: (data as any).total,
         };
@@ -379,17 +448,19 @@ const searchController = async (req: SaavnRequest, res: FastifyReply) => {
     }
   } catch (err) {
     console.error('Error in data mapping:', err);
+    return sendError(res, 'Data mapping error', err);
   }
 
-  res.code(code).send({ code, message, data: sanitizedData, error });
+  return sendSuccess(res, sanitizedData, message, 'saavn', code);
 };
 
 const artistController = async (req: SaavnRequest, res: FastifyReply) => {
   const { artistId } = req.params;
   const { page = 0, count = 50 } = req.query;
+  const languages = req.query.lang || req.query.languages;
   const url = `${config.saavn.baseUrl}`;
 
-  const { data, code, message, error } = await fetchGet(url, {
+  const { data, code, message, error } = await saavnFetch<any>(url, {
     params: {
       __call: config.saavn.endpoint.artist.link,
       token: artistId,
@@ -399,39 +470,51 @@ const artistController = async (req: SaavnRequest, res: FastifyReply) => {
       n_album: count,
       ...params,
     },
+    lang: languages,
   });
+
+  if (error) {
+    return sendError(res, message || 'Failed to fetch artist details', error, code);
+  }
+
   const extractedData = artistDataMapper(data);
-  res.code(200).send({ data: extractedData, code, message, error });
+  return sendSuccess(res, extractedData, message, 'saavn', code);
 };
 
 const songController = async (req: SaavnRequest, res: FastifyReply) => {
   try {
     const { songId } = req.params;
+    const languages = req.query.lang || req.query.languages;
     const url = `${config.saavn.baseUrl}`;
 
-    const { data, code, message, error } = await fetchGet(url, {
+    const { data, code, message, error } = await saavnFetch<any>(url, {
       params: {
         __call: config.saavn.endpoint.song.link,
         token: songId,
         type: 'song',
         ...params,
       },
+      lang: languages,
     });
 
-    console.log((data as any).songs[0].id);
+    if (error) {
+      return sendError(res, message || 'Failed to fetch song details', error, code);
+    }
 
-    const { data: lyricsData } = await fetchGet(url, {
+    const { data: lyricsData } = await saavnFetch<any>(url, {
       params: {
         __call: config.saavn.endpoint.get.lyrics,
         lyrics_id: (data as any)?.songs[0]?.id,
         ...params,
       },
+      lang: languages,
     });
+
     const songData = songsDetailsMapper(data);
-    const promiseArr = songData.modules.map((d) => {
+    const promiseArr = (songData.modules || []).map((d: any) => {
       return {
         title: d.heading,
-        promise: fetchGet(url, {
+        promise: saavnFetch<any>(url, {
           params: {
             __call: d.endpoint,
             song_id: d.songIds,
@@ -442,6 +525,7 @@ const songController = async (req: SaavnRequest, res: FastifyReply) => {
             pid: d.pid,
             ...params,
           },
+          lang: languages,
         }),
       };
     });
@@ -449,44 +533,39 @@ const songController = async (req: SaavnRequest, res: FastifyReply) => {
 
     const sections = results
       .map((result: any, index) =>
-        isValidArray(result.value.data)
+        isValidArray(result.value?.data)
           ? {
               heading: promiseArr[index].title,
-              data: isValidArray(result.value.data)
-                ? result.value.data.map((d) => albumDataMapper(d))
-                : undefined,
+              data: result.value.data.map((d: any) => {
+                if (d.type === 'song') return songDataSanitizer([d])[0];
+                if (d.type === 'album') return albumDataMapper(d);
+                if (d.type === 'playlist') return playlistDataMapper(d);
+                return albumDataMapper(d);
+              }),
+              source: 'saavn',
             }
           : null,
       )
       .filter((d: any) => d != null);
 
-    res.code(200).send({
-      data: {
-        song: songData.song,
-        sections: sections,
-        lyrics: (lyricsData as any)?.lyrics || '',
-      },
-      code,
-      message,
-      error,
-    });
+    const finalData = {
+      song: songData.song,
+      sections: sections,
+      lyrics: (lyricsData as any)?.lyrics || '',
+    };
+    return sendSuccess(res, finalData, message, 'saavn', code);
   } catch (error) {
-    res.code(400).send({
-      data: null,
-      code: 400,
-      message: 'failed to fetch',
-      error,
-    });
+    return sendError(res, 'Failed to fetch song details', error);
   }
 };
 
 const recommendedSongsController = async (req: SaavnRequest, res: FastifyReply) => {
   try {
     const { songId } = req.params;
-    const languages = req.query.languages;
+    const languages = req.query.lang || req.query.languages;
     const url = `${config.saavn.baseUrl}`;
 
-    const { data, code, message, error } = await fetchGet(url, {
+    const { data, code, message, error } = await saavnFetch<any>(url, {
       params: {
         __call: config.saavn.endpoint.song.recommended,
         pid: songId,
@@ -494,16 +573,17 @@ const recommendedSongsController = async (req: SaavnRequest, res: FastifyReply) 
         ...params,
         ctx: 'wap6dot0',
       },
+      lang: languages,
     });
-    const sanitizedData = songDataSanitizer(data);
-    res.code(code).send({ code, message, data: sanitizedData, error });
+
+    if (error) {
+      return sendError(res, message || 'Failed to fetch recommendations', error, code);
+    }
+
+    const sanitizedData = songDataSanitizer(data as any);
+    return sendSuccess(res, sanitizedData, message, 'saavn', code);
   } catch (error) {
-    res.code(400).send({
-      data: null,
-      code: 400,
-      message: 'failed to fetch',
-      error,
-    });
+    return sendError(res, 'Failed to fetch recommendations', error);
   }
 };
 
